gtk2: use improved assets for the GTK2 version
This commit improves a lot the assets we use to render the GTK2 version
of Adwaita, and makes them more consistent with the GTK3 version.
It also fixes some theming bugs in GIMP, and adds vertical and
horizontal padding to buttons, comboboxes, entries and spinbuttons, so
that they are not tied together as reported e.g. in

https://bugzilla.gnome.org/show_bug.cgi?id=682441