adwaita: use outline properties for focus instead of engine properties
This is a bit of a hack, but the problem is that using custom
engine properties for focus triggers some sort of slow path in the
GTK+ style validation, and this makes state changes sluggish in complex
treeview layouts.
Until focus rendering in GTK is fixed, the workaround is to use some
real CSS properties, which get a different caching treatment (but the
engine still manually queries the final value and does the rendering
itself).
This also gives us the occasion to get rid of those custom properties
altogether.

https://bugzilla.gnome.org/show_bug.cgi?id=684499