adwaita: override menu frame style when not running under the Shell
When we're not running under the shell, we're not guaranteed to have
shadows rendered around menus; this is especially the case for fallback
mode.
Since we don't really have a way of loading a different theme for
fallback mode from the settings daemon (or there might not even be a
settings daemon around in case we're running on a completely different
environment), we have to detect the presence of the Shell ourselves.
Install a GTK module for this (both for GTK2 and GTK3), that watches the
WM name on the GdkScreen and dynamically injects or removes frames on
menus when the WM name matches GNOME Shell.

https://bugzilla.gnome.org/show_bug.cgi?id=660271