assets: don't repeat background images for checks/radios
These images are used both for GtkCheckButton (which fortunately has an
indicator-size style property) and for GtkCellRendererToggle; in the
latter case, indicator-size is a property that can be changed from code
with a GObject property.
Since we used fixed-size assets for those, we must specify not to repeat
them, or it will look bad when the code changes indicator-size on a cell
renderer.

https://bugzilla.gnome.org/show_bug.cgi?id=685921