gtk2: make it possible to override GtkEntry's backgrounds
Some GTK2 applications like to override GtkEntry's background with their
own color for e.g. showing an error state.
Since we currently draw an asset for the entry's background, this
doesn't work.
To support this use case, always use the background color from the style
with no overrides, and put an overlay on top of it to get the same
gradient we used to get.
In addition, entry borders cannot be transparent anymore now, so we have
to provide two versions - a normal one and one that works on notebooks,
to avoid the base shining through the corners.

This also bumps the GTK2 required version to 2.24.15, which has a fix
that makes this possible.

https://bugzilla.gnome.org/show_bug.cgi?id=685712