gtk2: fix white border around background when overridden
The border assets we used previously had a fully opaque background
filled inside the border stroke.
When the background color is overridden by the application, part of the
frame will be drawn on top of the overridden background, which won't
look good if background colors don't match.
To fix this, use borders with a transparent fill in the middle.
Don't use this new assets for Libreoffice though, as it won't work there
for missing alpha channel support.

https://bugzilla.gnome.org/show_bug.cgi?id=685712