gtk2: fix regression in Chrome/Chromium toolbar drawing
When using the native GTK theme, Chromium uses an offscreen entry and
gtk_style_attach() to draw its background on the GtkHBox of the location
entry.
This won't take into account the base color of the GtkEntry itself, so
we have to special case the offscreen entry and use a solid background
for it.

https://bugzilla.gnome.org/show_bug.cgi?id=694661