Be more careful when overriding titlebar fonts
The current theme makes any text bold that appears anywhere
in a headerbar (including entries, buttons, etc). To fix that,
GTK+ now adds 'title' and 'subtitle' style classes to the
two labels. This commit makes use of those style classes to
restrict the font changes to just where it is desired.

https://bugzilla.gnome.org/show_bug.cgi?id=706154