gtk3: fix propagation for insensitive colors
Since 30609ba766dd055f41002dcb813225d619b1d951 we stopped using
wildcard selectors to propagate insensitive state colors. Turns out that
part of the commit was wrong, as typically the insensitive state is set
on a widget down the hierarchy, and not the toplevel, so relying on
inheritance from .background won't work.

This commit partially reverts some of the changes from
30609ba766dd055f41002dcb813225d619b1d951 to make insensitive colors work
correctly again.

https://bugzilla.gnome.org/show_bug.cgi?id=706009