Restrict menubutton styling
For menu buttons, we create an active state that makes the button
appear attached to the menu. That does not look so hot when the
button is actually popping up something else. GTK+ introduced
a style class, 'menu-button', to differentiate the case of a
button with a menu. Use it.

https://bugzilla.gnome.org/show_bug.cgi?id=723878