Updated old bits-list, fixed render script and rerendered assets.
For some reason rerendering the assets changed them, but it needed to be
done.

Looks like the assets in the svg are named differently than they are
used, producing lots of untracked assets on a rerender. I removed them
for the time being, I'll fix it properly with the new svg.

Also, the progressbar name was deleted accidentally in the old svg,
fixed that.